requirements.txt
flask
stripe
python-dotenv
requests
gunicorn
